Data.NutMeg

Description

Read Binary NutMeg Data

Data Types

type NutMeg = [(String, Plot)]

A NutMeg file consists of a list of Plots [(Plotname, Plot@)] Plotnames do not have to be unique in a NutMeg file. Data may be lost by turning this into a Map.

type Plot = Map String Wave

A Plot inside a NutMeg file consists of uniquely identified waveforms: [(Variable Name, Waveform)]

data Wave

Wrapper around Real or Complex valued Vector, so they can be in the same List.

Constructors

RealWave !(Vector Double)
Real valued (Double) wave form

ComplexWave !(Vector (Complex Double))
Complex valued ('Complex Double') wave form

data Flag

Flag indicating whether a plot is real or complex valued

Constructors

Real
Real valued (Double) plot

Complex
Complex valued ('Complex Double') plot

data Field

Data Field identifier in NutMeg (binary) Files

Constructors

Title
Title of the NutMeg file, usually first line of netlist

Date
Date of creation

Plotname
Name of a Plot

Flags
Whether the Plot is Real or Complex

NoVariables
No. Variables in the Plot

NoPoints
No. Points per Variable

Variables
String representation of Variables in the Plot

Binary
Binary Data, Double or 'Complex Double' encoded in Big Endian

Read raw binary data

readFile :: FilePath -> IO NutMeg

Read a binary nutmeg .raw file

readFile' :: FilePath -> IO NutMeg

Strict version of readFile

slurp :: ByteString -> NutMeg

Read all Plots encountered in a ByteString

Parsing NutMeg binary data

extractPlot :: ByteString -> ((String, Plot), ByteString)

Read The first plot encountered in ByteString String: ((Plotname, Plot), Remianing ByteString)

extractPlot'

Arguments

:: Flag
Real or Complex Data

-> Int
No. Variables

-> Int
No. Points

-> ByteString
Binary Data

-> [Wave]
Wave forms

Extract the wave forms from binary data given header information

parseHeader :: [ByteString] -> ((String, Flag, Int, Int), [String])

Extract Plot header information: ((Plotname, Flag, No. Variables, No. Points), [Variable Names])

readField :: Field -> ByteString -> String

Read a NutMeg field from a ByteString

Accessing Plot data

asVector :: Unbox a => Wave -> Vector a

Convert Waveform to unboxed Vector

flattenPlots :: [String] -> NutMeg -> Plot

Concatenate the Waves of a given list of Plot names. This will only work if the keys line up.

flattenPlots' :: NutMeg -> Plot

Concatenate waves of all Plots in NutMeg. Probably won't work as intended when variable names of the Plots don't line up

Utilities

r2c :: Unbox a => Int -> Int -> Vector a -> [Vector a]

Transpose Rows / Columns

concat :: Wave -> Wave -> Wave

Joins two Waves of the same type: wave1 ++ wave2 Attempting to concatenate a Real and Complex wave will result in an error