Data.NutMeg

Documentation

data NutMeg

Representation of NutMeg file contents, where plot names are mapped to corresponding plot types and data.

Constructors

NutMeg

Fields

nutTitle :: String

nutDate :: String

nutPlots :: !(Map String NutPlot)

data NutPlot

A plot as it is represented within the NutMeg File format.

Constructors

NutPlot

Fields

nutPlotName :: String

nutNumVars :: Int

nutNumPoints :: Int

nutVariables :: [String]

nutPlotType :: NutPlotType

nutData :: !(Map String NutWave)

data NutWave

NutMeg wave form data dypes

Constructors

NutRealWave !(Vector Double)

NutComplexWave !(Vector (Complex Double))

data NutPlotType

Real and Complex NutMeg plots

Constructors

NutRealPlot

NutComplexPlot

data NutField

Data Fields in NutMeg Files

Constructors

NutTitle

NutDate

NutPlotname

NutFlags

NutNoVariables

NutNoPoints

NutVariables

NutBinData

nutFieldName :: NutField -> ByteString

Translate fieldname to ByteString representation in NutMeg file.

nutFieldName' :: NutField -> ByteString

Convenient Transformer to Lazy ByteString

bytesPerReal :: Int

How many bytes per real data point

bytesPerReal' :: Int64

Convenient transformation to Int64 for certain calculations

bytesPerComplex :: Int

How many bytes per complex data point

bytesPerComplex' :: Int64

Convenient transformation to Int64 for certain calculations

popNutElement :: NutField -> ByteString -> (Maybe ByteString, ByteString)

Get the first NutMeg Element and the Rest

readNutRealRow :: Int64 -> Int -> ByteString -> Vector Double

Read a row of real valued data

readNutComplexRow :: Int64 -> Int -> ByteString -> Vector (Complex Double)

Read a row of complex valued data

nutPlot :: String -> NutPlotType -> Int -> Int -> [String] -> ByteString -> NutPlot

Construct NutPlot type

nutFlag :: String -> NutPlotType

Convert a flag from file to plot type

nutWave :: NutWave -> Either (Vector (Complex Double)) (Vector Double)

nutRealWave :: NutWave -> Maybe (Vector Double)

asRealVector :: NutWave -> Vector Double

flattenRealPlots :: [NutPlot] -> NutPlot

flattenComplexPlots :: [NutPlot] -> NutPlot

parseNutPlot :: ByteString -> NutPlot

Returns a NutPlot for a given ByteString segment.

parseNutMeg :: ByteString -> NutMeg

Returns a NutMeg for given file contents.

parseNutHeader :: ByteString -> [String]

Extract NutMeg header information

readNutRaw :: FilePath -> IO ByteString

Convenience function for reading a NutMeg File