Data.NutMeg

Description

Read Binary Nutmeg Data

Documentation

data NutMeg

Representation of NutMeg file contents, where plot names are mapped to corresponding plot types and data.

Constructors

NutMeg

Fields

nutTitle :: !String

nutDate :: !String

nutPlots :: !(Map String NutPlot)

data NutPlot

Convenient Transformer to Lazy ByteString nutFieldName' :: NutField -> BL.ByteString nutFieldName' = BL.fromStrict . nutFieldName

A plot as it is represented within the NutMeg File format.

Constructors

NutPlot

Fields

nutPlotName :: !String

nutNumVars :: !Int

nutNumPoints :: !Int

nutVariables :: ![String]

nutPlotType :: !NutPlotType

nutData :: !(Map String NutWave)

data NutWave

NutMeg wave form data dypes

Constructors

NutRealWave !(Vector Double)

NutComplexWave !(Vector (Complex Double))

data NutPlotType

Real and Complex NutMeg plots

Constructors

NutRealPlot

NutComplexPlot

data NutField

Data Fields in NutMeg Files

Constructors

NutTitle

NutDate

NutPlotname

NutFlags

NutNoVariables

NutNoPoints

NutVariables

NutBinData

nutFieldName :: NutField -> ByteString

Translate fieldname to ByteString representation in NutMeg file.

bytesPerReal :: Int64

Convenient transformation to Int64 for certain calculations

bytesPerComplex :: Int64

How many bytes per complex data point

nutPlot :: String -> NutPlotType -> Int -> Int -> [String] -> ByteString -> NutPlot

Construct NutPlot type

nutFlag :: String -> NutPlotType

Convert a flag from file to plot type

nutWave :: NutWave -> Either (Vector (Complex Double)) (Vector Double)

Convert a nut wave to either real or complex valued Vector

nutRealWave :: NutWave -> Maybe (Vector Double)

Convert a nut wave to a Real valued Vector

asRealVector :: NutWave -> Vector Double

Convert a nut wave to a Real valued Vector

readRealBinary :: ByteString -> [Double]

Read block of real data

readComplexBinary :: ByteString -> [Complex Double]

Read block of complex data

concatComplexWaves :: [NutWave] -> NutWave

Concatenate complex nut waves

joinComplexWaves :: NutWave -> NutWave -> NutWave

Convenience for zipping complex waves

nutComplexWave :: NutWave -> Maybe (Vector (Complex Double))

Convert to Complex valued Vector

asComplexVector :: NutWave -> Vector (Complex Double)

Convert to Complex valued Vector

flattenRealPlots :: [NutPlot] -> NutPlot

Concatenate Data of plots with same variable names

flattenComplexPlots :: [NutPlot] -> NutPlot

Flatten plots with same variable names

parseNutPlot :: [ByteString] -> NutPlot

Returns a NutPlot for a given ByteString segment.

parseNutPlots :: [ByteString] -> [NutPlot]

parseNutMeg :: ByteString -> NutMeg

Parse Nutmeg content

parseNutMeg' :: [ByteString] -> (NutMeg, Int64)

Parse nutmeg content with offset and return new offset

readNutRaw :: FilePath -> IO ByteString

Convenience function for reading a NutMeg File

readNutRaw' :: FilePath -> IO NutMeg